requirements.txt
requests
radon
lizard
python-dotenv
google-genai
